requirements.txt
yfinance
kafka-python
pandas
streamlit